requirements.txt
Django==2.1.2
django-money==0.14.3
djangorestframework==3.9.0
Pillow==5.3.0
py-moneyed==0.7.0
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
sorl-thumbnail==12.4.1
